make.sh: Set parallelism to number of CPU cores
Setting it to something higher than the number of physical CPU cores was
a good idea when we used to have slow magnetic storage. That way, at
least there was always one process waiting for IO.

With modern fast flash-based storage, this does not hold any more since
it is fast enough that we don't need to have a couple of processes ready
to wait.

It will probably have made work for the scheduler more challenging since
more processes were ready and processes were moved around processors.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>