rng-tools: update to 6.12
Please refer to https://github.com/nhorman/rng-tools/releases for a full
list of meaningful changes between 6.4 and 6.12. "--without-pkcs11" is
necessary to avoid additional dependencies, which do not make sense on
IPFire since there are no use-cases for it.

Rootfiles did not change, our patch to use RDRDAND on i586 as well is
still valid.

Signed-off-by: Peter Müller <peter.mueller@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>