firewall.cgi: Fix multiple usage of configured net or interface.
When configuring a standard network as source or target the same
interface would be pre-selected as firewall interface when editing an
existing rule.

In case an existing input rule with an configured firewall interface
should be changed, the same network device has been pre-selected in the
standard networks dropdown box.

This easily confuses users and may lead to false configurations when
saving an edited rule.

Signed-off-by: Stefan Schantl <stefan.schantl@ipfire.org>
Signed-off-by: Arne Fitzenreiter <arne_f@ipfire.org>