udev: Enable ntuple offloading feature in supported NICs
We are using CPU-affinity and packet steering functions in various
places in IPFire, but packets might still be received on a random CPU
core.

This feature enables that packets that belong to the same connection
(i.e. have the save tuple) will be steered to the same queue. This will
increase cache locality and decrease locking which results in higher
throughput.

https://www.kernel.org/doc/Documentation/networking/scaling.txt

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>
Acked-by: Peter Müller <peter.mueller@ipfire.org>
Signed-off-by: Arne Fitzenreiter <arne_f@ipfire.org>