nano: Update to 6.1
For details see:
https://www.nano-editor.org/news.php

"The behavior of ^K at a prompt has been enhanced: when there
is text after the cursor, just this text is erased. (In the usual
situation, however, when the cursor is at the end of the answer,
the behavior is as before: the whole answer is erased.)

At a prompt, M-6 copies the current answer into the cutbuffer.

Large external pastes into nano are handled more quickly."

Signed-off-by: Matthias Fischer <matthias.fischer@ipfire.org>
Reviewed-by: Peter Müller <peter.mueller@ipfire.org>