squid: Update from 5.2 => 5.4.1
For details see:
http://www.squid-cache.org/Versions/v5/changesets/SQUID_5_4_1.html

This is 'squid 5.4.1', containing the previous patch for Bug #5055.

Prior to this patch I reverted my previous patches 'squid: Update 5.2 => 5.4" and
'squid 5.4: Latest patch - Bug #5055 - from upstream' and  marked them as
'superseded' in patchwork.

For a better overview the 'squid-gcc11'-patch has been renamed again and moved
to an own squid-patch-directory.

Signed-off-by: Matthias Fischer <matthias.fischer@ipfire.org>
Reviewed-by: Peter Müller <peter.mueller@ipfire.org>