installer: Remove support for ReiserFS
https://lore.kernel.org/lkml/YhIwUEpymVzmytdp@casper.infradead.org/

ReiserFS is an old file system which has not been actively maintained in
the Linux kernel for a long time. It is potentially going to be removed
soon which is why we shouldn't encourage people to create new
installations with ReiserFS any more.

This patch removes support for ReiserFS from the installer.

We should keep it enabled in the kernel for as long as it is available,
but we will have to encourage users potentially to re-install on a
different file system.

Since ReiserFS isn't very popular any more, I don't think that there
will be many users left.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>
Acked-by: Peter Müller <peter.mueller@ipfire.org>
Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>